[Feature #112512027] Fixed failing test

+++ database/migrations/2016_01_27_180657_add_table_cost_and_label_to_restaurants_table.php
@@ -14,7 +14,7 @@ class AddTableCostAndLabelToRestaurantsTable extends Migration
     {
         Schema::table('tables', function (Blueprint $table) {
             $table->string('label')->nullable();
-            $table->decimal('cost')->nullable();
+            $table->double('cost')->nullable();
         });
     }
